+++ src/main/resources/db/migration/V2__changed_duration_type_for_Duration.sql
@@ -1,0 +1,5 @@
+ALTER TABLE workout
+    DROP COLUMN duration_min;
+
+ALTER TABLE workout
+    ADD duration_min BIGINT;